Delay switch to BlackCoin's ProtocolV2 for 1 year
Since BlackCoin’s ProtocolV2 is unproven, it will not be implemented at
launch (though it will remain in the source code and will be maintained
in the future).  A switch is presently set at the 1 year mark, but this
can be changed in the future if Protocol V2 proves either a success or
failure.  Regardless, this will need to be watched.